Remove dependency on riscv target file
Depending on this target JSON file resulted in picking up a lot of things
that aren't needed for the zkvm (i.e. the the CRS instructions for atomic
operations). Remove this dependency. This change brings in the baremetal
build tag to resolve several errors during compilation but it might not be
necessary for us to include this build tag.