add a new target: zkvm (#1)
* add a new target: zkvm

This is a first attempt to add the zkvm as a new target. At this point, all
files in testdata compile except for those involving goroutines and
channels. Some tests can run on r0vm but most fail. The failures require
further investigation and it's possible that this code being added contains mistakes...

* Update ELF file and keep the .init section at the top of the text section

TinyGo has header called `.init`. This is where the entry point lies. Move this
to the top of the .text section so that all elf files for the zkvm have
consistent locations for entry points relative to the start of the .text section.

* runtime: initialize .bss section

* Compile with -march=rv32im for riscv32 target

Not all riscv32 architectures want the c and a features.

* Add emulator command

Assumes that r0vm is installed on the path.

* Implement putchar

* temporary hack: picolibc: change flag so that picolibc emits correct code

If this flag is not set, the object files contain, floating point instructions/registers and compressed instructions...

* Remove dependency on riscv target file

Depending on this target JSON file resulted in picking up a lot of things
that aren't needed for the zkvm (i.e. the the CRS instructions for atomic
operations). Remove this dependency. This change brings in the baremetal
build tag to resolve several errors during compilation but it might not be
necessary for us to include this build tag.

* zkvm: configure scheduler

Scheduler uses the same context switch code as the tinygo.riscv. We can do this
because tinygo.riscv uses the same registers as riscv zkvm.

* zkvm: fix ticks() function and increase default stacksize

* picolibc: add option to use obsolete math double

This is needed for the zkvm execution environment which doesn't use floating-point operators or registers.
Enabling this option requires a few more files to pull in the exp2 function.

* RISC-V: differentiate between rv32im and rv32imac

The llvm-target triple starts with riscv32- does not mean that the board supports rv32imac.
Add logic to distinguish between RV32IM and RV32IMAC.

* zkvm: update target json